+++ pandas/core/index.py
@@ -1,3 +1,3 @@
 # flake8: noqa
-from pandas.indexes.api import *
-from pandas.indexes.multi import _sparsify
+from pandas.core.indexes.api import *
+from pandas.core.indexes.multi import _sparsify
