+++ pandas/core/indexes/frozen.py
