+++ pandas/tseries/period.py
@@ -37,14 +37,14 @@ from pandas._libs.period import (Period, IncompatibleFrequency,
                                  _quarter_to_myear)
 
 from pandas.core.base import _shared_docs
-from pandas.indexes.base import _index_shared_docs, _ensure_index
+from pandas.core.indexes.base import _index_shared_docs, _ensure_index
 
 from pandas import compat
 from pandas.util.decorators import (Appender, Substitution, cache_readonly,
                                     deprecate_kwarg)
 from pandas.compat import zip, u
 
-import pandas.indexes.base as ibase
+import pandas.core.indexes.base as ibase
 _index_doc_kwargs = dict(ibase._index_doc_kwargs)
 _index_doc_kwargs.update(
     dict(target_klass='PeriodIndex or list of Periods'))
