+++ setup.py
@@ -640,9 +640,9 @@ setup(name=DISTNAME,
                 'pandas.compat.numpy',
                 'pandas.core',
                 'pandas.core.dtypes',
+                'pandas.core.indexes',
                 'pandas.core.computation',
                 'pandas.core.sparse',
-                'pandas.indexes',
                 'pandas.errors',
                 'pandas.io',
                 'pandas.io.json',
